+++ .travis.yml
@@ -1,0 +1,11 @@
+language: python
+python:
+  -3.6
+install:
+  - pip install -r requirements.txt
+services:
+  -postgresql
+before script:
+  - psql -c 'create database mydb;' -U postgres
+# script:
+#   - pytest app/tests --cov=app/views --cov-report term-missing
